Add template specific JS bundles
