+++ package.json
@@ -13,10 +13,11 @@
     "url": "https://github.com/Shopify/starter-theme.git"
   },
   "devDependencies": {
-    "@shopify/slate-tools": "@alpha",
+    "@shopify/slate-tools": ">=1.0.0-alpha.27",
     "babel-preset-shopify": "^16.2.0",
     "eslint-plugin-shopify": "^19.0.0",
-    "stylelint-config-shopify": "^4.0.0"
+    "stylelint-config-prettier": "^3.0.4",
+    "stylelint-config-shopify": "^5.0.0"
   },
   "dependencies": {
     "@shopify/theme-a11y": "^1.0.0-alpha.3",
