+++ src/assets/scripts/templates/product.js
@@ -1,0 +1,1 @@
+// import '../sections/product';
