+++ src/assets/styles/theme.scss.liquid
@@ -1,0 +1,4 @@
+// Insert your styles with liquid below.
+// Import additional stylesheets into this sheet using CSS imports:
+//
+// @import './settings/variables';
